docker-compose.yml: using `ia_common` as the dep image instead of `ia_eisbase` for ia_video_common
Fixed the issue of build failing for VI/VA services on a fresh system due to
the image dependency issue. Using `ia_common` as the dep image instead of
`ia_eisbase` for `ia_video_common` service fixes it. Doing it this way to
avoid using ia_common, ia_video_common and ia_openvino_base in the dependency
list for VI/VA services. With this change, we can just use ia_openvino_base
as the dependency image for VI/VA services.

Signed-off-by: Vinod K B <vinod.k.b@intel.com>