video-common: Fixed encoding issue
The issue occurs when different encoding type is used
in VI and VA.

In udf manager if the encoding type (read from the
meta-data which is added in VI) doesn't match with
the encoding type (read in VA constructor and passed
to udf manager during initialization) them the
set-encoding function called in udf manager had issue
while reading from the meta-data.

There was no `encoding` key added in metadata. Only the
`encoding-type` and `encoding-level` is added.

Signed-off-by: sandeshk1 <sandesh.kumar.s@intel.com>