sample_realsense: Sending colorized Depth Frame
The behaviour of this UDF is chnaged ina way that now this UDF
will send a colorized depth frame instead of RGB frame.

The visualizer will show it till we alter visulizer to show
multi-frame senario. Additionally in order to see RGB frame
in visualizer from realsense camera user need to change the
UDF to dummy one.

Signed-off-by: Lalatendu das <lalatendu.das@intel.com>